Weather App/weather-alert-api/requirements.txt
fastapi[standard]
requests
python-dotenv
